fix: store the original handler in the pdb handler instead of globally
Storing the original SIGUSR2 handler globally causes an exception to
be raised on Windows. Instead, store the original handler on the
`_Handler` object. An added benefit is that restoring the original
handler doesn't default to the handler on SIGUSR2 at the time the
module was imported, but instead the handler on SIGUSR2 at the time
that `listen` is called.