fix: raise a warning on Windows
Windows is not supported. Instead of an exception being raised by
accessing SIGUSR2 which doesn't exist on Windows, do nothing and raise
a warning. This way, pdb-attach doesn't impact user code and users
don't have to do try-except imports.

Storing the original SIGUSR2 handler globally causes an exception to
be raised on Windows. Instead, store the original handler on the
`_Handler` object. An added benefit is that restoring the original
handler doesn't default to the handler on SIGUSR2 at the time the
module was imported, but instead the handler on SIGUSR2 at the time
that `listen` is called.

Running on GitHub Actions, pytest doesn't pick up the warning caused
by importing pdb_attach on Windows, but ironically displays the
warning. Skip trying to test the warning on import and just test those
on the `listen` and `unlisten` functions which pytest doesn't seem to
have a problem with.

Python 3.3 and 3.4 on Windows cannot find a compatible version of pytest
and so are skipped in the GitHub workflow. This isn't really a big deal since
pdb-attach doesn't work on Windows anyway and we just want to ensure the
tests for `PdbDetach` pass on Windows in general, which they do for versions
2.7 and 3.5+.

Related #4